requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.4.3
discord.py==1.5.0
idna==2.9
isort==4.3.21
joblib==0.16.0
lazy-object-proxy==1.4.3
mathparse==0.1.2
mccabe==0.6.1
multidict==4.7.6
nltk==3.5
packaging==20.4
Pint==0.16.1
pylint==2.5.3
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2020.1
PyYAML==3.13
regex==2020.6.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.2.19
textblob==0.15.3
toml==0.10.1
tqdm==4.47.0
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
